Logging enhancements (#72)
* Fix ContentFileVerifier not logging value

* Add logging to SteamClient

* Add logging to ContentDownloader + remove useless public methods

* Introduce Serilog with file logging

* Serialize objects & lists in logging

* Add serilog log levels settings to appsettings

* Fix SteamClientTests

* ContentDownloader small refactor

* Comment old logging configuration

* SteamClientTests refactor

* Rename Finally<T>() to TapOnBoth<T>()

This is to avoid confusion with other Finally methods which are not void

* Add tests for TapOnBoth()

* Enhance keys preparing logging

* Adjust keys preparing log levels

* Upgrade logging in Server/Modset Config

* Fix external keys not found bug in keys preparer

* ServerProvider logging enhancements

* Enhance logging in Server creation

* Fix not found keys not logged

* Enhance logging in KeysFinder and Preparer

* Fix keys copied counting

* Improve logging in content verification

* Add information log when mods update finishes

* Fix test not compiling

* Improve download progress logging

* Remove old console logging

* Add logging to HangfireManager

* Fix server status creation and logging

* Fix log message template in process discoverer

+++ ArmaForces.ArmaServerManager/Features/Steam/Content/ContentFileVerifier.cs
@@ -3,6 +3,7 @@ using System.IO;
 using System.IO.Abstractions;
 using System.Linq;
 using System.Security.Cryptography;
+using ArmaForces.Arma.Server.Extensions;
 using ArmaForces.ArmaServerManager.Features.Steam.Content.DTOs;
 using BytexDigital.Steam.ContentDelivery.Enumerations;
 using BytexDigital.Steam.ContentDelivery.Models;
@@ -29,7 +30,7 @@ namespace ArmaForces.ArmaServerManager.Features.Steam.Content
                 return Result.Success(contentItem);
             }
 
-            _logger.LogInformation("Item {contentItemId} doesn't have a directory.");
+            _logger.LogInformation("Item {ContentItemId} doesn't have a directory", contentItem.Id);
             return Result.Failure<ContentItem>("Item not exists.");
         }
 
@@ -52,6 +53,10 @@ namespace ArmaForces.ArmaServerManager.Features.Steam.Content
                 .Where(x => !expectedFiles.Contains(x))
                 .ToList();
 
+            if (redundantFiles.IsEmpty()) return;
+            
+            _logger.LogDebug("Found {Count} redundant files in {Directory}", redundantFiles.Count, directory);
+                
             foreach (var file in redundantFiles)
             {
                 _fileSystem.File.Delete(file);
